References #3, added initial implementation based on Sigma.js
It has quire a few flaws and gotchas. Also tuned the remaining code a
bit (every time tab is changes a Resize event is being fired; this deals
with situations of window resize while viewing some other tab and then
sets the charts to the right dimentions). Ajax loader indicator is also
missing.